Extract CategoryContainer
